+++ CHANGES.md
@@ -28,6 +28,7 @@
 -   Upgrade pat-calendar to use latest fullcalendar version (5.3.0).
 -   pat calendar: Add fullcalendar list views.
 -   pat calendar: Store the current date and view in query parameters.
+-   pat calendar: Fetch events from the backend.
 -   pat tooltip: Use tippy v6 based implementation.
 -   Allow overriding the public path from outside via the definition of a ``window.__patternslib_public_path__`` global variable.